requirements.txt
pdf2docx
streamlit
pdfplumber
python-pptx
pdf2image
